--- .npmignore
@@ -1,10 +1,0 @@
-.storybook
-config
-node_modules
-public
-scripts
-src
-.gitignore
-.travis.yml
-CONTRIBUTING
-
