+++ .travis.yml
@@ -8,17 +8,17 @@ notifications:
   email: false
 
 node_js:
-  - node
+  - 'node'
 
 script:
   - npm run test
   - npm run build:lib
-  - npm run clean:packagejson
 
 after_success:
-  - test $TRAVIS_BRANCH = "master" && npm run semantic-release
+  - npm i -g semantic-release
+  - cd lib
+  - npm run semantic-release
 
 branches:
-  except:
-    - /^v\d+\.\d+\.\d+$/
-
+  only:
+    - master
